Format power.cpp

+++ power.cpp
@@ -1,17 +1,18 @@
 ll power(ll x, ll y)
 {
-    ll res = 1;
- 
+    int res = 1;
+
     x = x % mod;
-  
-    if (x == 0) return 0;
- 
+
+    if (x == 0)
+        return 0;
+
     while (y > 0)
     {
         if (y & 1ll)
-            res = (res*x) % mod;
-        y = y>>1ll;
-        x = (x*x) % mod;
+            res = (res * x) % mod;
+        y = y >> 1ll;
+        x = (x * x) % mod;
     }
     return res;
 }
